requirements.txt
# python 3.9

tornado==6.1.0
six==1.16.0
Pillow==8.3.2
PyMySQL==1.0.2
dbutils==3.0.3
redis==3.5.3
pycparser==2.20
requests==2.24.0
configparser==5.0.2

numpy==1.25.0
pandas==2.0.2
pytz==2023.3
tzdata==2023.3

et-xmlfile==1.1.0
openpyxl==3.0.9
xltpl==0.18
JPype1==1.4.1

contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1

aliyun-python-sdk-core==2.13.36
bcrypt==3.2.0
PyJWT==1.7.1
log4py==2.2
pydbclib==2.2.3
sqlparse==0.4.4

pyasn1==0.5.0
rsa==4.9

apscheduler==3.10.4

elasticsearch==7.12.1

mysql-replication==0.30
requests-toolbelt==1.0.0
pyodbc==5.1.0
